move shit out of ecs.
