+++ cardinal/codec/codec.go
