+++ cardinal/component/component.go
