--- cardinal/ecs/archetype/archetype.go
@@ -1,1 +1,0 @@
-package archetype
