+++ cardinal/ecs/persona.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"fmt"
 
-	"pkg.world.dev/world-engine/cardinal/ecs/filter"
+	"pkg.world.dev/world-engine/cardinal/filter"
 	"pkg.world.dev/world-engine/cardinal/public"
 )
