+++ cardinal/ecs/persona_test.go
@@ -11,8 +11,8 @@ import (
 	"gotest.tools/v3/assert"
 
 	"pkg.world.dev/world-engine/cardinal/ecs"
-	"pkg.world.dev/world-engine/cardinal/ecs/filter"
-	"pkg.world.dev/world-engine/cardinal/ecs/inmem"
+	"pkg.world.dev/world-engine/cardinal/filter"
+	"pkg.world.dev/world-engine/cardinal/inmem"
 )
 
 func TestCreatePersonaTransactionAutomaticallyCreated(t *testing.T) {
