+++ cardinal/entity/entity.go
