+++ cardinal/filter/exact.go
