+++ cardinal/filter/not.go
