+++ cardinal/transaction/transaction_test.go
@@ -11,7 +11,7 @@ import (
 	"pkg.world.dev/world-engine/cardinal/ecs"
 	"pkg.world.dev/world-engine/cardinal/public"
 
-	"pkg.world.dev/world-engine/cardinal/ecs/inmem"
+	"pkg.world.dev/world-engine/cardinal/inmem"
 )
 
 type ScoreComponent struct {
